Big refactoring (#51)
* Big refactoring WIP

* Fixed warnings

For some reason (likely due to new headers for simd in Clay) lots of
more symbols gets included which causes warnings for __u128 not being
FFI safe. We don't need any of these and blocks various header files
doesn't seem to work, but I will give it another go

* cargo fmt

* Fixed all the tests

* Fixed no-std

* Cargo fmt

* Managed to block away some extra symbols

* Removed layout tests to reduce code-size

* Make sure to inline setters

* Removed clippy supression not needed anymore

+++ src/errors.rs
@@ -1,7 +1,7 @@
 use crate::bindings::*;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u32)]
+#[repr(u8)]
 pub enum ErrorType {
     /// Thrown if the text measurement function is never provided to clay and you try using
     /// `Clay::text`
@@ -29,7 +29,7 @@ pub struct Error<'a> {
 impl From<Clay_ErrorData> for Error<'_> {
     fn from(value: Clay_ErrorData) -> Self {
         Self {
-            type_: unsafe { core::mem::transmute::<u32, ErrorType>(value.errorType) },
+            type_: unsafe { core::mem::transmute::<u8, ErrorType>(value.errorType) },
             text: value.errorText.into(),
         }
     }
